+++ salt/files/ServiceConnectionEvolution/scripts/collect.py
@@ -1,0 +1,2 @@
+from serconevo.agent import service_collect_agent
+service_collect_agent.main()