+++ salt/salt_run.sh
@@ -1,0 +1,1 @@
+nohup salt '*' state.apply SCE.agent --out=yaml --out-file-append --out-file=/tmp/agent.log.`date +%F` &