+++ serconevo/agent/service_collect_agent.py
@@ -42,7 +42,7 @@ usage = '''
     --version, -V           输出版本号。
 ''' % sys.argv[0]
 
-identify_line = "=*=" * 30
+identify_line = "=*=" * 20
 
 db_con = db_connect.DbInitConnect()
 config_parser = db_con.python_config_parser
@@ -125,16 +125,25 @@ def get_server_ip():
     ip_set = set()
     nia = psutil.net_if_addrs()
     pLogger.debug(nia)
-    for v in nia.values():
-        if v[0].address != '127.0.0.1':
-            pLogger.debug('get ip {}'.format(v[0].address))
-            ip_set.add(v[0].address)
+    for i_face in nia.values():
+        pLogger.debug("i_face is {}".format(i_face))
+        for addr in i_face:
+            pLogger.debug("addr is {}".format(addr))
+            if addr.address != '127.0.0.1' and addr.address != '::1':
+                if str(addr.family) == 'AddressFamily.AF_INET' or str(addr.family) == 'AddressFamily.AF_INET6':
+                    pLogger.debug('get ip {}'.format(addr.address))
+                    ip_set.add(addr.address)
+                else:
+                    pLogger.warn("{!r} is not a INET address".format(addr))
+            else:
+                pLogger.warn("{!r} is loop address ".format(addr))
     pLogger.debug("ip_set = {!r}".format(ip_set))
     return ip_set
 
 
 def ps_collect():
     server_ip = get_server_ip()
+    listen_ip_list = server_ip
     server_uuid = get_server_uuid()
     # 直接使用process_iter()迭代实例化每个进程.
     try:
@@ -142,6 +151,7 @@ def ps_collect():
             pLogger.debug("{}\nPorcesses [{}] begin to process.".format(identify_line, process.pid))
             # 判断进程是否还在运行.
             if process.is_running():
+                pid = process.pid
                 # 一次性抓取运行快照
                 process_listen_port = set()
                 process_connection_ip_port = set()
@@ -151,7 +161,6 @@ def ps_collect():
                     if connections:
                         # 只处理有连接的进程
                         name = process.name()
-                        pid = process.pid
                         exe = process.exe()
                         cwd = process.cwd()
                         cmdline = process.cmdline()
@@ -164,41 +173,73 @@ def ps_collect():
                         # num_ctx_switches = process.num_ctx_switches()
                         username = process.username()
                         # memory_full_info = process.memory_full_info()
-                
+
+                        # LISTENS
                         for connection in connections:
                             # 监听连接单独处理
                             if connection.status == psutil.CONN_LISTEN:
                                 # 如果连接地址为127.0.0.1,替换成非loop地址
+                                pLogger.debug("CONN_LISTEN connection is {!r}".format(connection))
                                 process_listen_port.add(connection.laddr[1])
+                                pLogger.debug("connection.laddr is {!r}".format(connection.laddr))
                                 # none LISTENS process
-                
-                        for connection in connections:
-                            if connection.status != psutil.CONN_LISTEN and \
-                                            connection.status != psutil.CONN_NONE and \
-                                            connection.laddr[1] not in process_listen_port:
-                                pLogger.debug("connection {!r} has connection.raddr is {!r}".format(connection,
-                                                                                                    connection.raddr))
-                                process_connection_ip_port.add(connection.raddr)
-                
+                            else:
+                                pLogger.warn("{!r} is not LISTEN connection!".format(connection))
                         # take out from set.
                         if process_listen_port:
                             pLogger.debug("{!r} process_listen_port is {!r}".format(pid, process_listen_port))
-                            pLogger.info("{} insert listen to DB.".format(pid))
-                            import2db(service_listens_table, server_ip, process_listen_port, name, pid, exe, cwd,
-                                      cmdline, status, create_time, username, server_uuid)
+                            pLogger.info("{!r} insert listen to DB.".format(pid))
+                            for l_port in process_listen_port:
+                                for l_ip in listen_ip_list:
+                                    import2db(service_listens_table, l_ip, l_port, name, pid, exe, cwd,
+                                              cmdline, status, create_time, username, server_uuid)
                         else:
                             pLogger.warn("Process {} with pid {} has no listen ports!".format(process, pid))
-                        # 从laddr中排除自身监听的端口
+
+                        # CONNECTIONS
+                        for connection in connections:
+                            # 从连接池中排除连接到自身监听的端口的链接
+                            if connection.status != psutil.CONN_LISTEN \
+                                    and connection.status != psutil.CONN_NONE \
+                                    and connection.laddr[1] not in process_listen_port:
+                                pLogger.debug("[{!r}] connection {!r} has connection.raddr is {!r}".format(
+                                        pid,
+                                        connection,
+                                        connection.raddr))
+                                process_connection_ip_port.add(connection.raddr)
+                            else:
+                                pLogger.warn("{!r} is not our collect connection!".format(connection))
+
                         if process_connection_ip_port:
+                            db_insert_ip_list = []
                             pLogger.debug("{!r} process_connection_ip_port is {!r}".format(pid,
                                                                                            process_connection_ip_port))
                             for connection_raddr in process_connection_ip_port:
                                 c_ip, c_port = connection_raddr
-                                pLogger.info("{} insert connection to DB.".format(pid))
-                                import2db(service_connections_table, c_ip, c_port, name, pid, exe, cwd,
-                                          cmdline, status, create_time, username, server_uuid, local_ip=server_ip)
+                                pLogger.debug("Before convert c_ip is {!r}".format(c_ip))
+                                # ip to ipv4
+                                c_ip = c_ip.split(":")[-1]
+                                pLogger.debug('c_ip, c_port is {} : {}'.format(c_ip, c_port))
+                                # 如果是连接127.0.0.1或::1,替换成本机IP
+                                if c_ip == '1' or c_ip == '127.0.0.1':
+                                    pLogger.debug("c_ip {!r} is loop, replace local real ip.".format(c_ip))
+                                    for c_ip_local_to_real in listen_ip_list:
+                                        pLogger.info("{} insert connection {!r} to DB insert list."
+                                                     .format(pid, c_ip_local_to_real))
+                                        db_insert_ip_list.append((c_ip_local_to_real, c_port))
+                                else:
+                                    pLogger.debug("c_ip {} insert into db insert list.".format(c_ip))
+                                    db_insert_ip_list.append((c_ip, c_port))
+
+                            for insert in db_insert_ip_list:
+                                insert_ip, insert_port = insert
+                                import2db(service_connections_table, insert_ip, insert_port, name, pid, exe,
+                                          cwd, cmdline, status, create_time, username,
+                                          server_uuid, local_ip=server_ip)
                         else:
-                            pLogger.warn("{} has no connections.".format(pid))
+                            pLogger.warn("process_connection_ip_port is empty!")
+                    else:
+                        pLogger.warn("{} has no connections.".format(pid))
     
             else:
                 pLogger.warn("process {} is already not exist!".format(process.pid))
@@ -215,7 +256,7 @@ def create_list_to_str(item, num):
 @db_commit
 def import2db(table, ip, port, p_name, p_pid, p_exe, p_cwd, p_cmdline, p_status, p_create_time, p_username,
               server_uuid, **kwargs):
-    pLogger.debug(kwargs)
+    pLogger.debug("import2db kwargs is {!r}".format(kwargs))
     # 1 column_values create
     ip_column = ""
     port_column = ""
@@ -231,12 +272,14 @@ def import2db(table, ip, port, p_name, p_pid, p_exe, p_cwd, p_cmdline, p_status,
             ip_column = "l_ip"
             port_column = "l_port"
             columns_num = 11
+            pLogger.debug("Table is {}, column_num is {}".format(service_listens_table, columns_num))
             # format_str = create_list_to_str('{!r}', columns_num)
             # format_str_value = create_list_to_str('"{!r}"', columns_num)
         elif table == service_connections_table and kwargs['local_ip']:
             ip_column = "c_ip"
             port_column = "c_port"
             columns_num = 12
+            pLogger.debug("Table is {}, column_num is {}, append column.".format(service_listens_table, columns_num))
             columns_list_process.append('local_ip')
             values_list.append(kwargs['local_ip'])
     except ValueError: